initial v2.46

+++ Source/7zip/7zip/Compress/LZ/LZInWindow.h
@@ -4,7 +4,7 @@
  * This file is a part of LZMA compression module for NSIS.
  * 
  * Original LZMA SDK Copyright (C) 1999-2006 Igor Pavlov
- * Modifications Copyright (C) 2003-2022 Amir Szekely <kichik@netvision.net.il>
+ * Modifications Copyright (C) 2003-2006 Amir Szekely <kichik@netvision.net.il>
  * 
  * Licensed under the Common Public License version 1.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -92,9 +92,9 @@ public:
     _streamPos -= subValue;
   }
 
-  bool NeedMove(size_t numCheckBytes)
+  bool NeedMove(UInt32 numCheckBytes)
   {
-    size_t reserv = _pointerToLastSafePosition - (_buffer + _pos);
+    UInt32 reserv = _pointerToLastSafePosition - (_buffer + _pos);
     return (reserv <= numCheckBytes);
   }
 };
